+++ pysui/sui/sui_builders/get_builders.py
@@ -10,6 +10,7 @@
 #    limitations under the License.
 
 # -*- coding: utf-8 -*-
+# pylint: disable=line-too-long
 
 """Sui Builders: Simple sui_getXXX calls."""
 
@@ -20,6 +21,8 @@ from pysui.sui.sui_types.scalars import SuiString, SuiInteger, ObjectID, SuiNull
 from pysui.sui.sui_types.collections import SuiMap, EventID
 from pysui.sui.sui_types.address import SuiAddress
 from pysui.sui.sui_txresults.single_tx import (
+    DynamicFields,
+    SuiCoinBalance,
     SuiCoinMetadata,
     CoinBalances,
     SuiCoinObjects,
@@ -63,22 +66,103 @@ class GetCoinMetaData(_NativeTransactionBuilder):
         return [self.coin_type]
 
 
-class GetCoinTypeBalance(_NativeTransactionBuilder):
-    """GetCoinTypeBalance Return the total coin balance for each coin type."""
+class GetAllCoinBalances(_NativeTransactionBuilder):
+    """GetAllCoinBalances Returns the total coin balances, for all coin types, owned by the address owner.."""
 
-    def __init__(self, *, owner: SuiAddress, coin_type: SuiString):
+    def __init__(self, *, owner: SuiAddress):
+        """__init__ Initializes builder with address to fetch coin balances for.
+
+        :param owner: the owner's Sui address
+        :type owner: SuiAddress
+        """
+        super().__init__("sui_getAllBalances", handler_cls=CoinBalances, handler_func="ingest_data")
+        # super().__init__("sui_getAllBalances")
+        self.owner = None
+        self.set_owner(owner)
+
+    def set_owner(self, owner: SuiAddress) -> "GetAllCoinBalances":
+        """Set the owner property."""
+        if isinstance(owner, SuiAddress):
+            self.owner = owner
+            return self
+        raise ValueError(f"{owner} is not of type SuiAddress")
+
+    def _collect_parameters(self) -> list[SuiBaseType]:
+        """_collect_parameters Returns expected RPC parameters."""
+        return [self.owner]
+
+
+class GetAllCoins(_NativeTransactionBuilder):
+    """GetAllCoins Returns all Coin objects owned by an address."""
+
+    def __init__(
+        self,
+        *,
+        owner: SuiAddress,
+        limit: SuiInteger = None,
+        cursor: ObjectID = None,
+    ):
+        """__init__ Initialize builder with address and optional page limits and cursor.
+
+        :param owner: the coin owner's Sui address
+        :type owner: SuiAddress
+        :param limit: maximum number of items per page, defaults to None
+        :type limit: SuiInteger, optional
+        :param cursor: Optional ObjectID as the starting item in returned page, defaults to None
+        :type cursor: ObjectID, optional
+        """
+        super().__init__("sui_getAllCoins", handler_cls=SuiCoinObjects, handler_func="from_dict")
+        self.set_owner(owner)
+        self.limit = SuiNullType()
+        self.cursor = SuiNullType()
+        if limit:
+            self.set_limit(limit)
+        self.cursor = cursor if cursor else SuiNullType()
+
+    def set_owner(self, owner: SuiAddress) -> "GetCoinTypeBalance":
+        """Set the owner property."""
+        if isinstance(owner, SuiAddress):
+            self.owner = owner
+            return self
+        raise ValueError(f"{owner} is not of type SuiAddress")
+
+    def set_limit(self, limit: Union[int, SuiInteger]) -> "GetCoins":
+        """Sets the maximum values returned on the result page."""
+        if isinstance(limit, SuiInteger):
+            self.limit = limit
+            return self
+        if isinstance(limit, int):
+            self.limit = SuiInteger(limit)
+            return self
+        raise ValueError(f"{limit} is not of type SuiInteger")
+
+    def set_cursor(self, cursor: ObjectID) -> "GetCoins":
+        """Optional ObjectID (of sample ``coin_type``)."""
+        self.cursor = cursor
+        return self
+
+    def _collect_parameters(self) -> list[SuiBaseType]:
+        """_collect_parameters Returns expected RPC parameters."""
+        return [self.owner, self.cursor, self.limit]
+
+
+class GetCoinTypeBalance(_NativeTransactionBuilder):
+    """GetCoinTypeBalance Return the total coin balance for a coin type."""
+
+    def __init__(self, *, owner: SuiAddress, coin_type: SuiString = None):
         """__init__ Initializes builder with address and coin type to fetch balances for.
 
         :param owner: the owner's Sui address
         :type owner: SuiAddress
-        :param coin_type: fully qualified type names for the coin (e.g., 0x2::sui::SUI)
-        :type coin_type: SuiString
+        :param coin_type: fully qualified type names for the coin (e.g., 0x2::sui::SUI), defaults to None
+        :type coin_type: SuiString, optional
         """
-        super().__init__("sui_getBalance", handler_cls=CoinBalances, handler_func="ingest_data")
+        super().__init__("sui_getBalance", handler_cls=SuiCoinBalance, handler_func="from_dict")
         self.owner = None
-        self.coin_type = None
+        self.coin_type = SuiNullType()
         self.set_owner(owner)
-        self.set_coin_type(coin_type)
+        if coin_type:
+            self.set_coin_type(coin_type)
 
     def set_owner(self, owner: SuiAddress) -> "GetCoinTypeBalance":
         """Set the owner property."""
@@ -103,32 +187,37 @@ class GetCoinTypeBalance(_NativeTransactionBuilder):
 
 
 class GetCoins(_NativeTransactionBuilder):
-    """Return the list of Coin objects owned by an address."""
+    """Return the list of Coin objects of specific coin_type owned by an address."""
 
     def __init__(
         self,
         *,
         owner: SuiAddress,
-        coin_type: SuiString,
-        limit: SuiInteger,
+        coin_type: SuiString = None,
+        limit: SuiInteger = None,
         cursor: ObjectID = None,
     ):
-        """__init__ Initialize builder with address,coin_type, page limits and cursor.
+        """__init__ Initialize builder.
 
         :param owner: the coin owner's Sui address
         :type owner: SuiAddress
-        :param coin_type: fully qualified type names for the coin (e.g., 0x2::sui::SUI)
-        :type coin_type: SuiString
-        :param limit: maximum number of items per page
-        :type limit: SuiInteger
+        :param coin_type: fully qualified type names for the coin (e.g., 0x2::sui::SUI), defaults to None
+        :type coin_type: SuiString, optional
+        :param limit: maximum number of items per page, defaults to None
+        :type limit: SuiInteger, optional
         :param cursor: Optional ObjectID as the starting item in returned page, defaults to None
         :type cursor: ObjectID, optional
         """
         super().__init__("sui_getCoins", handler_cls=SuiCoinObjects, handler_func="from_dict")
-        self.set_limit(limit)
-        self.cursor = cursor if cursor else SuiNullType()
         self.set_owner(owner)
-        self.set_coin_type(coin_type)
+        self.coin_type = SuiNullType()
+        self.limit = SuiNullType()
+        self.cursor = SuiNullType()
+        if coin_type:
+            self.set_coin_type(coin_type)
+        if limit:
+            self.set_limit(limit)
+        self.cursor = cursor if cursor else SuiNullType()
 
     def set_owner(self, owner: SuiAddress) -> "GetCoinTypeBalance":
         """Set the owner property."""
@@ -265,6 +354,70 @@ class GetObjectsOwnedByObject(_NativeTransactionBuilder):
     def _collect_parameters(self) -> list[ObjectID]:
         """Collect the call parameters."""
         return [self.object_id]
+
+
+class GetDynamicFieldObject(_NativeTransactionBuilder):
+    """GetDynamicFieldObject when executed, return the dynamic field object information for a specified object."""
+
+    def __init__(self, parent_object_id: ObjectID, field_name: SuiString) -> None:
+        """__init__ Builder initializer.
+
+        :param parent_object_id: The ID of the queried parent object
+        :type parent_object_id: ObjectID
+        :param field_name: The Name of the dynamic field
+        :type field_name: SuiString
+        """
+        super().__init__("sui_getDynamicFieldObject", handler_cls=ObjectRead, handler_func="factory")
+        self.parent_object_id = parent_object_id
+        self.name = field_name
+
+    def _collect_parameters(self) -> list[ObjectID]:
+        """Collect the call parameters."""
+        return [self.parent_object_id, self.name]
+
+
+class GetDynamicFields(_NativeTransactionBuilder):
+    """GetDynamicFields when executed, returns the list of dynamic field objects owned by an object."""
+
+    def __init__(
+        self, parent_object_id: ObjectID, limit: Union[int, SuiInteger] = None, cursor: Union[str, ObjectID] = None
+    ) -> None:
+        """__init__ Builder initializer.
+
+        :param parent_object_id: The ID of the queried parent object
+        :type parent_object_id: ObjectID
+        :param limit: Maximum item returned per page, default to [QUERY_MAX_RESULT_LIMIT] if not specified, defaults to None
+        :type limit: Union[int, SuiInteger], optional
+        :param cursor: Optional paging cursor, defaults to None
+        :type cursor: Union[str, ObjectID], optional
+        """
+        super().__init__("sui_getDynamicFields", handler_cls=DynamicFields, handler_func="from_dict")
+        self.parent_object_id = parent_object_id
+        self.cursor = SuiNullType()
+        self.limit = SuiNullType()
+        if cursor:
+            self.set_cursor(cursor)
+        if limit:
+            self.set_limit(limit)
+
+    def set_cursor(self, cursor: ObjectID) -> "GetDynamicFields":
+        """Optional paging cursor."""
+        self.cursor = cursor
+        return self
+
+    def set_limit(self, limit: Union[int, SuiInteger]) -> "GetDynamicFields":
+        """Sets the maximum values returned on the result page."""
+        if isinstance(limit, SuiInteger):
+            self.limit = limit
+            return self
+        if isinstance(limit, int):
+            self.limit = SuiInteger(limit)
+            return self
+        raise ValueError(f"{limit} is not of type SuiInteger")
+
+    def _collect_parameters(self) -> list[ObjectID]:
+        """Collect the call parameters."""
+        return [self.parent_object_id, self.cursor, self.limit]
 
 
 class GetObject(_NativeTransactionBuilder):
@@ -467,11 +620,18 @@ class GetCommittee(_NativeTransactionBuilder):
         :type epoch: SuiInteger, optional
         """
         super().__init__("sui_getCommitteeInfo", handler_cls=CommitteeInfo, handler_func="factory")
-        self.epoch: SuiInteger = epoch if epoch else SuiInteger(None)
+        self.epoch = SuiNullType()
+        if epoch:
+            self.set_epoch(epoch)
 
     def set_epoch(self, epoch: SuiInteger) -> "GetCommittee":
         """Set epoch."""
-        self.epoch = epoch
+        if isinstance(epoch, int):
+            self.epoch = SuiInteger(epoch)
+        elif isinstance(epoch, SuiInteger):
+            self.epoch = epoch
+        else:
+            raise ValueError(f"epoch: {epoch} is not int or SuiInteger")
         return self
 
     def _collect_parameters(self) -> list[SuiBaseType]:
@@ -572,20 +732,21 @@ class GetEvents(_NativeTransactionBuilder):
     def __init__(
         self,
         *,
-        query: Union[SuiString, SuiMap] = None,
+        query: Union[SuiString, SuiMap],
         cursor: EventID = None,
-        limit: Union[SuiInteger, SuiNullType] = None,
+        limit: SuiInteger = None,
         descending_order: SuiBoolean = None,
     ) -> None:
         """Initialize builder."""
         inargs = locals().copy()
         super().__init__("sui_getEvents", handler_cls=EventQueryEnvelope, handler_func="from_dict")
-        self.query: Union[SuiString, SuiMap] = None
-        self.cursor: EventID = None
-        self.limit: Union[SuiInteger, SuiNullType] = None
-        self.descending_order: SuiBoolean = False
-        for hit in self.events_kwords & set(inargs.keys()):
-            setattr(self, hit, inargs[hit])
+        self.query: Union[SuiString, SuiMap]
+        for key, value in inargs.items():
+            if key in self.events_kwords:
+                if value:
+                    setattr(self, key, inargs[key])
+                else:
+                    setattr(self, key, SuiNullType())
 
     def _collect_parameters(self) -> list[SuiBaseType]:
         """Collect the call parameters."""
@@ -670,20 +831,20 @@ class GetTxs(_NativeTransactionBuilder):
     def __init__(
         self,
         *,
-        query: Union[str, SuiMap] = None,
-        curser: str = None,
+        query: Union[str, SuiMap],
+        cursor: SuiString = None,
         limit: SuiInteger = None,
         descending_order: SuiBoolean = None,
     ) -> None:
         """Initialize builder."""
         inargs = locals().copy()
         super().__init__("sui_getTransactions", handler_cls=TransactionQueryEnvelope, handler_func="from_dict")
-        self.query: Union[str, SuiMap] = None
-        self.cursor: str = None
-        self.limit: SuiInteger = None
-        self.descending_order: SuiBoolean = False
-        for hit in self.txs_kwords & set(inargs.keys()):
-            setattr(self, hit, inargs[hit])
+        for key, value in inargs.items():
+            if key in self.txs_kwords:
+                if value:
+                    setattr(self, key, inargs[key])
+                else:
+                    setattr(self, key, SuiNullType())
 
     def _collect_parameters(self) -> list[SuiBaseType]:
         """Collect the call parameters."""
@@ -710,21 +871,21 @@ class GetTransactionsInRange(_NativeTransactionBuilder):
         """."""
         if isinstance(start, SuiInteger):
             self.start = start
-            return self
-        if isinstance(start, int):
+        elif isinstance(start, int):
             self.start = SuiInteger(start)
-            return self
-        raise ValueError(f"{start} is not an integer type.")
+        else:
+            raise ValueError(f"{start} is not an integer type.")
+        return self
 
     def set_end(self, end: SuiInteger) -> "GetTransactionsInRange":
         """."""
         if isinstance(end, SuiInteger):
             self.end = end
-            return self
-        if isinstance(end, int):
+        elif isinstance(end, int):
             self.end = SuiInteger(end)
-            return self
-        raise ValueError(f"{end} is not an integer type.")
+        else:
+            raise ValueError(f"{end} is not an integer type.")
+        return self
 
     def _collect_parameters(self) -> list[SuiBaseType]:
         """Collect the call parameters."""